% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{add_user_to_space}
\alias{add_user_to_space}
\title{Invite user to space}
\usage{
add_user_to_space(user_email, space_id, email_invite = TRUE,
  email_message = "You are invited to this space",
  space_role = "contributor")
}
\arguments{
\item{user_email}{Email address of the user to add}

\item{space_id}{ID number of the space}

\item{email_invite}{Indicates whether an email should be sent to the user with the invite, `TRUE` by default}

\item{email_message}{Message to be sent to the user should the `email_invite` flag be set to `TRUE`}

\item{space_role}{Desired role for the user in the space}
}
\description{
Invites a user to a space with their email address and with a given role,
and can also prompt an invitation email if `email_invite` is set to `TRUE`.
}
