% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaces.R
\name{get_spaces}
\alias{get_spaces}
\title{Return spaces that user has access to}
\usage{
get_spaces()
}
\description{
This will typically be the first call that a user makes to help determine
which spaces they have access to and to retrieve the space IDs that are
needed in many of the later calls for managing space memberships and
invitations.
}
