% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{initialize_token}
\alias{initialize_token}
\title{Initializes the token that will be used in all future requests.}
\usage{
initialize_token()
}
\description{
Initializes the token that will be used in all future requests.
}
