% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invitations.R
\name{invitation_get}
\alias{invitation_get}
\title{Retrieve list of outstanding invitations}
\usage{
invitation_get(space_id)
}
\arguments{
\item{space_id}{ID number of the space}
}
\description{
Retrieves the list of all outstanding invitations for the given space.
}
