% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invitations.R
\name{invitation_send}
\alias{invitation_send}
\alias{invitation_send.numeric}
\alias{invitation_send.tbl_invitation}
\title{Send or Resend Invitations}
\usage{
invitation_send(invitations)

\method{invitation_send}{numeric}(invitations)

\method{invitation_send}{tbl_invitation}(invitations)
}
\arguments{
\item{invitations}{An invitation ID number or a data frame
of invitations returned by \code{space_invitation_list()}.}
}
\description{
Sends or resends invitation with a given invitation ID that was previously
created in RStudio Cloud. Invitation IDs are unique across all spaces, hence
this function does not also depend on a space ID.
}
\seealso{
\code{\link[=space_member_add]{space_member_add()}}
}
