% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invitations.R
\name{invitations_for_space}
\alias{invitations_for_space}
\title{Retrieve list of outstanding invitations}
\usage{
invitations_for_space(space_id)
}
\arguments{
\item{space_id}{ID number of the space}
}
\description{
Retrieves the list of all outstanding invitations for the given space.
}
