% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{members_for_space}
\alias{members_for_space}
\title{Returns the list of members for this space.}
\usage{
members_for_space(space_id)
}
\arguments{
\item{space_id}{is the id of the space that the user has access to}
}
\description{
Returns the list of members for this space.
}
