% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaces.R
\name{project_get}
\alias{project_get}
\title{Return projects in space}
\usage{
project_get(space_id, filters = NULL)
}
\arguments{
\item{space_id}{ID number of the space}

\item{filters}{is a vector of filters to be AND'ed and applied to the request}
}
\description{
Returns the projects in a given space.
}
