% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaces.R
\name{projects_for_space}
\alias{projects_for_space}
\title{Return projects in space}
\usage{
projects_for_space(space_id)
}
\arguments{
\item{space_id}{ID number of the space}
}
\description{
Returns the projects in a given space.
}
