% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{refresh_rscloud_token}
\alias{refresh_rscloud_token}
\title{Convenience function to refresh the rscloud token since the refresh flow isn't implemented yet today.
As of May 28, 2019, the initial token that is granted is good for an hour.  If you receieve a 401 error,
try running the refresh command once again, and then try your original command again.}
\usage{
refresh_rscloud_token()
}
\description{
Convenience function to refresh the rscloud token since the refresh flow isn't implemented yet today.
As of May 28, 2019, the initial token that is granted is good for an hour.  If you receieve a 401 error,
try running the refresh command once again, and then try your original command again.
}
