% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{remove_user_from_space}
\alias{remove_user_from_space}
\title{Removes the given user_id from the space}
\usage{
remove_user_from_space(user_id, space_id)
}
\arguments{
\item{user_id}{the id of the user to be remoed from the space}

\item{space_id}{the id for the space to be modified}
}
\description{
Removes the given user_id from the space
}
