% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invitations.R
\name{rescind_invitation}
\alias{rescind_invitation}
\title{Cancels an existing invitation.}
\usage{
rescind_invitation(invitation_id)
}
\arguments{
\item{invitation_id}{is the id of an invitation that was previously created in the system.}
}
\description{
Cancels an existing invitation.
}
