% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaces.R
\name{roles_for_space}
\alias{roles_for_space}
\title{Return valid roles for space}
\usage{
roles_for_space(space_id)
}
\arguments{
\item{space_id}{ID number of the space}
}
\description{
Returns valid roles for a given space and the permissions associated with each role.
}
