% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaces.R
\name{rscloud_space}
\alias{rscloud_space}
\title{Construct a Space Object}
\usage{
rscloud_space(space_id = NULL, name = NULL)
}
\arguments{
\item{space_id}{The space ID.}

\item{name}{The space name.}
}
\value{
An \code{rscloud_space} object.
}
\description{
Returns a space object given the space ID or the space name.
}
\details{
Exactly one of \code{space_id} or \code{name} must be specified.
}
