% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaces.R
\name{rscloud_space_list}
\alias{rscloud_space_list}
\title{List Spaces}
\usage{
rscloud_space_list(filters = NULL)
}
\arguments{
\item{filters}{A vector of filters to be AND'ed and applied to the request.}
}
\value{
A data frame of space attributes.
}
\description{
This will typically be the first call that a user makes to help determine
which spaces they have access to and to retrieve the space IDs that are
needed in many of the later calls for managing space memberships and
invitations.
}
