% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invitations.R
\name{send_invitation}
\alias{send_invitation}
\title{Send or resend invitation}
\usage{
send_invitation(invitation_id)
}
\arguments{
\item{invitation_id}{ID number of the invitation}
}
\description{
Sends or resends invitation with a given invitation ID that was previously
created in RStudio Cloud. Invitation IDs are unique across all spaces, hence
this function does not also depend on a space ID.
}
