% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_token}
\alias{setup_token}
\title{This is called automatically on package load; you should only need to call yourself
if you've changed the setting in `config.yml`}
\usage{
setup_token()
}
\description{
This is called automatically on package load; you should only need to call yourself
if you've changed the setting in `config.yml`
}
\keyword{internal}
