% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaces.R
\name{space_info}
\alias{space_info}
\title{Get Space Information}
\usage{
space_info(space)
}
\arguments{
\item{space}{A space object created using \code{rscloud_space()}.}
}
\description{
Obtain information on a space.
}
