% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invitations.R
\name{space_invitation_list}
\alias{space_invitation_list}
\title{Retrieve list of outstanding invitations}
\usage{
space_invitation_list(space, filters = NULL)
}
\arguments{
\item{space}{A space object created using `rscloud_space()`.}

\item{filters}{A vector of filters to be AND'ed and applied to the request.}
}
\description{
Retrieves the list of all outstanding invitations for the given space.
}
