% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/members.R
\name{space_member_remove}
\alias{space_member_remove}
\alias{space_member_remove.numeric}
\alias{space_member_remove.character}
\alias{space_member_remove.data.frame}
\title{Remove Members}
\usage{
space_member_remove(space, users, content_action = NULL, ask = TRUE)

\method{space_member_remove}{numeric}(space, users, content_action = NULL, ask = TRUE)

\method{space_member_remove}{character}(space, users, content_action = NULL, ask = TRUE)

\method{space_member_remove}{data.frame}(space, users, content_action = NULL, ask = TRUE)
}
\arguments{
\item{space}{A space object created using \code{rscloud_space()}.}

\item{users}{ID number or email of the user to be removed, or a data frame
with either a \code{user_id} or \code{email} column.}

\item{content_action}{What to do with the users content after they are
removed from the space. Options for the content are to: \code{leave_in_space}
leaving the content where it is, \code{archive_in_space} moving the content to
the space archive, and \code{trash_in_space} moving the content to the spaces
trash.}

\item{ask}{Whether to ask user for confirmation of deletion.}
}
\description{
Removes members.
}
