% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{space_project_list}
\alias{space_project_list}
\title{List Projects}
\usage{
space_project_list(space, filters = NULL)
}
\arguments{
\item{space}{A space object created using `rscloud_space()`.}

\item{filters}{A vector of filters to be AND'ed and applied to the request.}
}
\description{
Returns the projects in a given space.
}
