% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{space_remove_user}
\alias{space_remove_user}
\title{Remove user from space}
\usage{
space_remove_user(space_id, user_id)
}
\arguments{
\item{space_id}{ID number of the space}

\item{user_id}{ID number of the user}
}
\description{
Removes a user with a given user ID from the space.
}
