% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaces.R
\name{space_role}
\alias{space_role}
\title{Return valid roles for space}
\usage{
space_role(space_id)
}
\arguments{
\item{space_id}{ID number of the space}
}
\description{
Returns valid roles for a given space and the permissions associated with each role.
}
