% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaces.R
\name{space_role_list}
\alias{space_role_list}
\title{Valid Roles for Space}
\usage{
space_role_list(space)
}
\arguments{
\item{space}{A space object created using \code{rscloud_space()}.}
}
\description{
Returns valid roles for a given space and the permissions associated with each role.
}
