% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{user_get}
\alias{user_get}
\title{Get list of members for space}
\usage{
user_get(space_id)
}
\arguments{
\item{space_id}{ID number of the space}
}
\description{
Returns the list of members for a given space. You must either be the admin
of the space or your role must have permissions to see the members list.
}
