% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{add_user_to_space}
\alias{add_user_to_space}
\title{Invite a user by email addrss to the space.}
\usage{
add_user_to_space(user_email, space_id,
  email_message = "You are invited to this space", email_invite = TRUE,
  space_role = "contributor", access_code = NULL)
}
\arguments{
\item{user_email}{email address of the user to add}

\item{email_message}{the message to be sent to the user should the email_invite flag be set to TRUE}

\item{email_invite}{indicates whether an email should be sent to the user with the invite}

\item{space_role}{the desired role for the user in the space}

\item{access_code}{TODO: No idea what this does, need to talk to the team about it :)}
}
\description{
Invite a user by email addrss to the space.
}
