% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaces.R
\name{get_spaces}
\alias{get_spaces}
\title{Returns the spaces that this user has access to.
This will typically be the first call that a user makes to help determine which spaces they have access to and
to retrieve the space_ids that aare needed in many of the later calls for managing space memberships and invitations.}
\usage{
get_spaces()
}
\description{
Returns the spaces that this user has access to.
This will typically be the first call that a user makes to help determine which spaces they have access to and
to retrieve the space_ids that aare needed in many of the later calls for managing space memberships and invitations.
}
\seealso{

}
