% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invitations.R
\name{invitation_rescind}
\alias{invitation_rescind}
\alias{invitation_rescind.numeric}
\alias{invitation_rescind.tbl_invitation}
\title{Cancel Invitations}
\usage{
invitation_rescind(invitations)

\method{invitation_rescind}{numeric}(invitations)

\method{invitation_rescind}{tbl_invitation}(invitations)
}
\arguments{
\item{invitations}{An invitation ID number or a data frame
of invitations returned by \code{space_invitation_list()}.}
}
\description{
Cancels existing invitations.
}
