% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invitations.R
\name{invitations_for_space}
\alias{invitations_for_space}
\title{Retrieves the list of all outstanding invitations for the given space}
\usage{
invitations_for_space(space_id)
}
\arguments{
\item{space_id}{id of the space we are interested in getting the invites for}
}
\description{
Retrieves the list of all outstanding invitations for the given space
}
