% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{members_for_space}
\alias{members_for_space}
\title{Get list of members for space}
\usage{
members_for_space(space_id)
}
\arguments{
\item{space_id}{ID number of the space}
}
\description{
Returns the list of members for a given space. You must either be the admin
of the space or your role must have permissions to see the members list.
}
