% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaces.R
\name{projects_for_space}
\alias{projects_for_space}
\title{Returns the projecst for this space}
\usage{
projects_for_space(space_id)
}
\arguments{
\item{space_id}{is an id of an existing space that the user has access to.}
}
\description{
Returns the projecst for this space
}
