% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{remove_user_from_space}
\alias{remove_user_from_space}
\title{Remove user from space}
\usage{
remove_user_from_space(user_id, space_id)
}
\arguments{
\item{user_id}{ID number of the user}

\item{space_id}{ID number of the space}
}
\description{
Removes a user with a given user ID from the space.
}
