% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invitations.R
\name{rescind_invitation}
\alias{rescind_invitation}
\title{Cancel existing invitation}
\usage{
rescind_invitation(invitation_id)
}
\arguments{
\item{invitation_id}{ID number of the invitation}
}
\description{
Cancels an existing invitation with a given invitation ID that was previously
created in RStudio Cloud. Invitation IDs are unique across all spaces, hence
this function does not also depend on a space ID.
}
