% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaces.R
\name{roles_for_space}
\alias{roles_for_space}
\title{Returns the valid roles available for this space.}
\usage{
roles_for_space(space_id)
}
\arguments{
\item{space_id}{is an id of an existing space that the user has access to.}
}
\description{
Returns the valid roles available for this space.
}
