% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rscloud-package.R
\docType{package}
\name{rscloud-package}
\alias{rscloud}
\alias{rscloud-package}
\title{rscloud: Provides an R client to the rstudio.cloud REST APIs}
\description{
Provides an R client to the rstudio.cloud REST APIs to facilitate the adding and removing of members to spaces.
}
\author{
\strong{Maintainer}: Tareef Kawaf \email{tareef@rstudio.com}

}
\keyword{internal}
