% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invitations.R
\name{send_invitation}
\alias{send_invitation}
\title{Sends or resends the invitation to the recipient}
\usage{
send_invitation(invitation_id)
}
\arguments{
\item{invitation_id}{is the id of an invitation that was previously created in the system.}
}
\description{
Sends or resends the invitation to the recipient
}
