% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{space_member_remove}
\alias{space_member_remove}
\title{Remove member from space}
\usage{
space_member_remove(space_id, user_id)
}
\arguments{
\item{space_id}{ID number of the space}

\item{user_id}{ID number of the user}
}
\description{
Removes a member with a given user ID from the space.
}
