% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{space_member_remove}
\alias{space_member_remove}
\alias{space_member_remove.numeric}
\alias{space_member_remove.character}
\alias{space_member_remove.data.frame}
\title{Remove Members}
\usage{
space_member_remove(space, users, remove_projects = NULL, ask = TRUE)

\method{space_member_remove}{numeric}(space, users, remove_projects = NULL, ask = TRUE)

\method{space_member_remove}{character}(space, users, remove_projects = NULL, ask = TRUE)

\method{space_member_remove}{data.frame}(space, users, remove_projects = NULL, ask = TRUE)
}
\arguments{
\item{space}{A space object created using `rscloud_space()`.}

\item{users}{ID number or email of the user to be removed, or a data frame
with either a `user_id` or `email` column.}

\item{remove_projects}{Whether to remove user's projects from the workspace
and move them to their personal space or to keep the projects in the
workspace.}

\item{ask}{Whether to ask user for confirmation of deletion.}
}
\description{
Removes members.
}
