% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/members.R
\name{space_member_usage}
\alias{space_member_usage}
\title{Get member usage information}
\usage{
space_member_usage(space, filters = NULL)
}
\arguments{
\item{space}{A space object created using \code{rscloud_space()}.}

\item{filters}{A vector of filters to be AND'ed and applied to the request.}
}
\description{
Returns the list of members and their usage information for a given space.
You must either be the admin of the space or your role must have permissions
to see the members list.
}
\examples{
\dontrun{
# Usage in the last 90 days for each user
space_member_usage(space, filters = list(groupby = "user_id", from = "90d"))
}

}
