% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaces.R
\name{space_project_get}
\alias{space_project_get}
\title{Return projects in space}
\usage{
space_project_get(space_id)
}
\arguments{
\item{space_id}{ID number of the space}
}
\description{
Returns the projects in a given space.
}
