% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaces.R
\name{space_project_get}
\alias{space_project_get}
\title{Return projects in space}
\usage{
space_project_get(space_id, filters = NULL)
}
\arguments{
\item{space_id}{ID number of the space}

\item{filters}{is a string that will be concatenated onto the filter call see the [lucid docs](https://buildmaster.rstudioservices.com/job/hostedapps/job/lucid-pipeline/job/master/API/api.html#header-how-to-filter)}
}
\description{
Returns the projects in a given space.
}
